✨ feat: add persist middleware to save in local storage [KOBG-11]
